Create gene expression specificity scores for scRNAseq data

Create gene expression specificity scores for scRNAseq data

Usage

create_ctd(
  matrix = NULL,
  meta_lev_1 = NULL,
  meta_lev_2 = NULL,
  outdir = NULL,
  mhc_genes = NULL,
  group_name = NULL,
  threads = NULL
)

Arguments

matrix

Raw count gene matrix - needs to be genes x cell and annotation data

meta_lev_1

A vector of cell IDs for level 1 cell types

meta_lev_2

A vector of cell IDs for level 1 cell types

outdir

The output directory to store the ctd object

mhc_genes

A set of MHC genes to be removed

group_name

An string identifier for the dataset

threads

Number of cores to assign to process. Default is NULL, which will auto detect total cores and estimate number to use.

Value

A CTD object containing

directory where ctd object was stored

Examples

data('seurat_small', 'mhc_genes')
create_ctd(seurat_small@assays$RNA@data,
                     meta_lev_1 = seurat_small$cluster_level_1,
                     meta_lev_2 = seurat_small$cluster_level_2,
                     outdir = 'test',
                     mhc_genes = mhc_genes,
                     group_name = 'brain_study',
                     threads = 1)
#> 
#> 
#> Creating CTD object ... 
#> MHC genes removed: 0
#> Loading required package: SeuratObject
#> Loading required package: sp
#> The legacy packages maptools, rgdal, and rgeos, underpinning the sp package,
#> which was just loaded, will retire in October 2023.
#> Please refer to R-spatial evolution reports for details, especially
#> https://r-spatial.org/r/2023/05/15/evolution4.html.
#> It may be desirable to make the sf package available;
#> package maintainers should consider adding sf to Suggests:.
#> The sp package is now running under evolution status 2
#>      (status 2 uses the sf package in place of rgdal)
#> 
#> Running CPM ...
#> 
#> Dropping uninformative genes cpm norm ...
#> Check 21310Check 1452
#> Warning: sctSpecies_origin not provided. Setting to 'mouse' by default.
#> 1 core(s) assigned as workers (1 reserved).
#> Converting to sparse matrix.
#> Checking for non-expressed genes.
#> -2918 / 18392 non-expressed genes dropped
#> Checking for cells with no expressed genes.
#> DGE:: Limma...
#> 8,495 / 18,392 genes dropped @ DGE adj_pval_thresh < 1e-05
#> Time difference of 3.894935 secs
#> 
#> Gene counts:
#> 
#> RAW:21310
#> RAW_NO_MHC:21310
#> CPM_DROP_GENES:9897
#> 
#> Creating CTD object ... 
#> 
#> 1 core(s) assigned as workers (1 reserved).
#> Converting to sparse matrix.
#> + Calculating normalized mean expression.
#> Converting to sparse matrix.
#> Converting to sparse matrix.
#> + Calculating normalized specificity.
#> Converting to sparse matrix.
#> Converting to sparse matrix.
#> Converting to sparse matrix.
#> Converting to sparse matrix.
#> Loading required namespace: ggdendro
#> + Saving results ==>  test/ctd_brain_study.rda
#> [1] "test/ctd_brain_study.rda"